Enable movements when running emacsclient on multiple sessions
When invoking a subshell (with `shell-command`) tmux will think that
it is running in the same session that started emacsclient, i.e:
1. Start tmux and emacsclient in the first session;
2. Create a new session, split the current window in two panes and
   connect to emacsclient;
3. From emacs (in session #2) try to move to the other pane (C-h, C-j,
   C-k, C-l);

You won't leave emacs' pane in this session, because `tmux
select-pane` was executed in the emacs pane from session #1.

Right now, some functions won't work as expected. ie: all functions
that rely on `{last}` variable from tmux, because I couldn't find a
way to get that value from the subshell. So, those functions will keep
running as they were invoked from the first session.