Add support for running commands/manipulating sibling pane
With the limitation that we can only do that when there are at most
two panes in current window (emacs' pane + another pane).

This was the only way I found that enable some behaviour similar to
using the `{last}` variable from tmux, but expanding to sessions that
emacsclient was not started on.